Cast offset to the value type expected by fprintf
Compiler was complaining about the incorrect value type:
fs_visualize.c: In function ‘main’:
fs_visualize.c:158:3: warning: format ‘%lld’ expects argument of type ‘long long int’, but argument 3 has type ‘off_t’ [-Wformat]

Type cast added, compiler is happy now.

+++ fs_visualize.c
@@ -155,7 +155,7 @@ int main(int argc, char **argv)
 	w = ceil(sqrt(pixels));
 
 	fprintf(stderr, "len %lld, pix_size %lu, pixels %d, w %d\n",
-		len, pix_size, pixels, w);
+		(long long int)len, pix_size, pixels, w);
 	img = malloc(w * w);
 	if (img == NULL) {
 		perror("");
